Add Keymap for triggering datatips via the keyboard
Summary:
**A previous version of this landed with just the keymaps**. The rest of this code is necessary for completing
the test plan and a decent experience with the keymap.

This adds an api to trigger datatips located under the editor cursor. With the addition of modifier datatyps, this means that before the hyperclick keychord (cmd-alt-enter) is completed with cmd-alt, the user receives a preview of where they will jump with the keyboard.

Additionally, users can simple hold `alt` to get a typehint of the symbol below whereever the cursor is posiitioned.

Reviewed By: jgebhardt

Differential Revision: D5106237

fbshipit-source-id: 60622a52b25ed905531b0e6e37a69e14dae2efa6