Debugger launch dialog does not show providers that are loaded late
Summary:
The debug provider store was not raising any events when the set of loaded providers changed. This results in some providers being missing since there is a race for which are loaded before the debugger package initializes and  caches the provider list for the launch/attach UX.

This change adds an emitted event and updates the provider list.

Reviewed By: johnislarry

Differential Revision: D5167957

fbshipit-source-id: 11b993783abe9c8c4418b7184513eea95e1fc7d0