handle 'telemetry/event'
Summary:
I've added handling for telemetry/event.

The main motivation is that I want the Hack server to be able to write stuff into the Nuclide logs, so that when they're collected in a bug dump, they'll be more informative.

The level of LSP-server tracing is controlled by the LSP parameter 'trace' which is passed to the initialize method. LSP doesn't specify how the LSP-server should respond to that. I figure that responding with telemetry events is a good way to do it!

Reviewed By: hansonw

Differential Revision: D5174265

fbshipit-source-id: 995f11ce30415d1812543f9f06616d3b8a5b4140