Scuba logging
Summary:
This adds Scuba logging of Start, Restart-on-crash and Exceptions.

(There's a little bit of redundancy: exception message + stacktrace are logged under 'lsp-exception' and also under 'lsp-start' if they occur then. I think that's desirable. The beginning is a very delicate time, and I'd like to see all start-related information in the same place in scuba.)

Reviewed By: hansonw

Differential Revision: D5185071

fbshipit-source-id: 260f155a45691c7d6c7f9f38e8553776cc810dd5