Filetree defaults to "uncommitted" not "head" changes
Summary:
We added a feature for the filetree to show "head / stack" changes in addition to the existing uncommitted changes.

We set the default during beta-testing to "head" so it would get more visibility. This diff changes the default to be "uncommitted" in preparation for wider rollout.

Scuba analysis: https://fburl.com/2nifar54

* When the feature first came out, more people picked "uncommitted" than "head". But we don't have the data to know how many people left it on the default of "head" without clicking any buttons.

* On an ongoing basis, the regular users of this feature end up toggling between Head/Uncommitted fairly routinely.

Reviewed By: mostafaeweda

Differential Revision: D5182952

fbshipit-source-id: b5c9847e1d36ea928f08b99f63847c483a1566a4