Remove Chrome dependency(28/N): Fix the LLDB breakpoint resolve notification error in new channel
Summary:
The design of LLDB engine may send breakpoint resolve event before setBreakpointByURL is complete. This race caused new channel's breakpoint manager state machine to violate -- we may get a breakpoint resolve before breakpoint id is assigned.
It is almost impossible to collate this breakpoint resolve with the original setBreakpointByURL in LLDB engine side so I relax the situation in the channel.

Reviewed By: ebluestein

Differential Revision: D5192217

fbshipit-source-id: f1eda5501a4ac75236e55ce6299eccfbd3f5518b