Omit locations from generated proxy types
Summary:
The locations in the generated proxy types are purely used for display in validation errors (OTOH the definition locations are used at registration time to check for duplicate names... another issue).

After looking into it, I think it's fine if we drop the type locations and use the nearest definition location (with improved messaging).

Reviewed By: zertosh

Differential Revision: D5196255

fbshipit-source-id: 8c588f4be077e2ee1372157414b9484b54f2d7f6