Refactor file parser to be more like a visitor
Summary:
The goal of this stack is to make the service parser lazy, i.e. only parse the types that are actually used in each service.

Currently, each FileParser stores the definitions exported by the file. Instead, let's move the definitions into the ServiceParser and have FileParser behave more like a visitor (which accumulates the definitions inside the ServiceParser).

This makes FileParser almost stateless (but not quite). The next diff finishes this work and globally caches file parsers.

Reviewed By: zertosh

Differential Revision: D5196253

fbshipit-source-id: d3caecb1a846a40552686b645d300aeada479418