Kill getDefinitionById in language services
Summary:
`getDefinitionById` was a half-implemented concept intended for use with Context View, where you could potentially link to other symbols by their definition.

I don't think anybody has any plans to continue that work, so this deletes all the code to support that. (Note that this means that `hh --get-definition-by-id` is now unused).

ljw1004 or anyone else please correct me if there's any plans to make use of this.

Reviewed By: ljw1004

Differential Revision: D5196635

fbshipit-source-id: 1036f0bb4086df53435f08fc76996bce0847849f