Fix format-on-save when closing unsaved editors
Summary:
If you close an unsaved file with `format` and get prompted to save (and confirm), currently you lose your data.

This is because Atom synchronously 1) prompts you 2) saves the file and then 3) destroys it, which works fine when all the I/O is synchronous.

However with format-on-save this is no longer the case. This works around the issue by listening for 'will-destroy' events and synchronously terminating all formatting logic when that happens, ensuring that the final `save()` is the real save.

Reviewed By: wbinnssmith

Differential Revision: D5206643

fbshipit-source-id: e794da7a13185b72f69b1d1c4164c8d45e574906