verbose tracing
Summary:
When I first wrote this code I'd misread vscode-jsonrpc and so was calling trace('verbose').

The vscode-jsonrpc API actually takes in a number from an enumeration. So this diff changes it.

NOTE: I'm really not sure the best way to handle enums. I've seen ways to do it that look pretty clunky (export type for the values, then export const for the enum map, then declare that the enum map obeys the type for the values). But that solution doesn't look like it's a lot of extra words for marginal benefit.

Reviewed By: hansonw

Differential Revision: D5190803

fbshipit-source-id: 2b48136e48cbf90d3b8cdd817279ac9bcbfaea5c