verbosity controller by GK
Summary:
With this diff, Nuclide's LSP connection will do verbose tracing only if so instructed by a GK.

I've initially set the GK to be true for absolutely everyone (i.e. everyone will get verbose logging).

In the past, the verbosity of the nuclide logs for Hack was a problem. This diff is a route to mitigate it.

We could also have more-or-less achieved this tweak by changing the log-level. But log-level is controlled by user in their configuration. I really wanted a GK, so we can roll out LSP and (say) always get verbose logs from beta-testers.

Reviewed By: hansonw

Differential Revision: D5191758

fbshipit-source-id: 59d0417bfe0a2f01a9620929009a65f01aeaa0c3