remove now-unneeded ENOENT check
Summary:
Previously if you did "await spawn(...).toPromise()" but the specified file didn't exist, then it would return a process whose PID field was null. We used to catch this case ourselves.

But a recent checkin changed the behavior of "spawn" so that now it will throw an error in that case:
https://github.com/facebook/nuclide/commit/51eec924a19eaf4208c14c3232c2dd1950ad838d

Therefore we can remove the code.

Reviewed By: matthewwithanm

Differential Revision: D5193768

fbshipit-source-id: 035e75a01f5587cf23e4951295b6366d10b2bbf8