Convert nuclide-context-view to use definition providers directly
Summary:
I'm getting rid of `nuclide-definition-service` - all it does is aggregate the definition providers together. Users can do this aggregation themselves with just a few lines, so it's not worth an extra package in atom-ide-ui.

I slightly modified the logic to actually remove an optimization (of avoiding the editor subscription when there's no definition service). This never really occurs in the context of Nuclide, so it's actually a premature optimization.

Reviewed By: matthewwithanm

Differential Revision: D5205800

fbshipit-source-id: fd276b8cffa2f5c0143f7f8ab676880deceed411