tighten up on stdout reporting
Summary:
It's been really useful that Nuclide reports messages from stderr and stdout in error situations when trying to spawn the LSP executable. e.g. printing messages like "--from option not supported".

But I noticed that in cases where the LSP server sends an LSP telemetry notification message and then closes, LspLanguageService was printing that stdout.

I think the true answer about stdout is: Nuclide should report messages from stdout up to the point where it's established that the process is speaking JsonRPC correctly. But once it gets to that point, then the contents of stdout are surely just a long complicated JsonRPC packet which isn't human-readable and which is presumably handled elsewhere inside LspLanguageService.

That's what this diff does.

Reviewed By: hansonw

Differential Revision: D5204682

fbshipit-source-id: 7e0a4e55b964ad386b1e52496636348160777ed3