Temporary workaround for debugger layout issue on Atom 1.17+
Summary:
Currently if the debugger spawns into a vertical dock that already has other pane(s) in it, it splits the dock into two columns and then spawns into its own column. But if another item (like Outline View or commit form) spawns in afterwards, the new item ends up as a tiny tab sharing a pane with one of the debugger elements, instead of getting the entire height of the column.

This is a temp fix to alleviate the problem until we can build a more robust layout engine on top of the workspace views API. If the debugger is present but not running, it will just get out of the way and hide itself when the new pane comes into the dock. If it is running, it will split the dock vertically and add itself back in next to the new pane.

Reviewed By: matthewwithanm

Differential Revision: D5215561

fbshipit-source-id: 81e05a952fc051ae26464ed1d37b920102f645f2