Create device type tasks
Summary:
These tasks are bound to a device type (e.g. android) and not a specific device.
As an example, I created a task that will be used to set a custom adb or sdb path, which is useful to keep the same adb version when using adb from both nuclide and terminal.

Reviewed By: sas

Differential Revision: D5197738

fbshipit-source-id: 2263a754073db2bf324c7a2ac175016a058f7bfc