Delete unused settings-pane-integration spec
Summary: This test is failing in Atom 1.17 due to the elements expected by `elementForKeyPath` not being available synchronously. Upon further investigation, it turns out that this spec has bitrotted into mostly no-ops, since most of the logic was switched off or commented out. The only thing this test is checking is whether our checkbox component works, which seems low-value.

Reviewed By: matthewwithanm

Differential Revision: D5222625

fbshipit-source-id: 6df50b134176170f368a8a84e6d7319eb65b7fa2