Several refactors
Summary:
D5213425 made me realize that we should get the adb (and the port) before each actual adb call. In fact, each rpc function can trigger several adb calls at different times, which is enough to have different adb paths (and ports).
I removed the AdbSdbBase class, which is very ugly and moved to a composition approach using a DebugBridge class, which is the one who makes the actual call to adb or sdb. The port should be used there as well.
I also create a config type which right now contains the path but will contain the port.

Reviewed By: ebluestein

Differential Revision: D5225219

fbshipit-source-id: 974d4760a21bcd0ffd484a38b93502a142e3990a