RFC: Show cwd in FileTree title
Summary:
Current working directory (cwd) deeply impacts a lot of the functionality in Nuclide in an implicit way.  Currently this shows up in the user interface only as a little star next to a node that can be scrolled off the screen in the File Tree.

This diff makes this information more prominent.  When cwd is set it shows the basename of the cwd (the last directory component, which is the same as the name of the node in the tree).  If the file is remote, it also shows the remote machine name.

Reviewed By: matthewwithanm

Differential Revision: D5222256

fbshipit-source-id: d83bc138dd1a1522d460b200fb05841fb16bb207