Auto-expand directories with one child only
Summary:
If a directory contains a single child, which is also a directory, when expanding the parent, the chances are, you want to expand the child as well. These are especially common in Java where directory structure mimics packages and often has long common prefixes.

This diff adds this functionality.

The behavior can be turned off with a setting.

Reviewed By: matthewwithanm

Differential Revision: D5220532

fbshipit-source-id: 8d88f8c2d81e3379a4d8fcd9e516d63a4179bd70