Some improvements in the poller
Summary:
Some subscribers were sending a variety of hosts for localhost (e.g. '', 'local', 'fbobjc/.../', etc) which breaks caching.
Now, as the cache key, i'm using '' for all the local hosts.
Also, I created a new package with a Cache class that I was using for the device panel which is kind of cute and I'm reusing it in the poller

Reviewed By: sas

Differential Revision: D5231778

fbshipit-source-id: 3d9944e0a671e8ae7907754f6576e4b4cc9e5194