Fix fallback (Windows) fuzzy filename matcher
Summary:
The Windows fallback matcher behaves rather bizarrely - it appears to not match any uppercase character in your original search.

I'm not sure what the original intent was, but it makes more sense to me to make both the needle and haystack lowercase and only compare after that.

There's some separate handling for camel-case matching, so I think this is more correct.

Reviewed By: ebluestein

Differential Revision: D5250011

fbshipit-source-id: 6e6d5d044a7fdf19670b911bf06e6c51d8a5da2c