Flag some incorrect file URLs
Summary:
A file URL has the form
  file://localhost/PATH
  file://HOST/PATH
  file:///PATH
  file:/PATH
The path uses / as segment separators, and (after UTF8 encoding) and uses percent-escaping for everything outside the list of safe characters `[a-zA-Z0-9-._~!$&'()*+,;=:@]`.

It's not the same as just prefixing a path with the string 'file://' !!!

I've added some notes where we're not doing it right.

Reviewed By: ebluestein

Differential Revision: D5238443

fbshipit-source-id: 58b7c537514995836f206e4ddf6e6acb70efc0a6