Restructure file diffs to use git diff
Summary:
Instead of hitting hg diff for each time a file diff needs to be computed, stash the file contents at the current head and use git diff to compute the diff.

This sees an approximately 3x speedup in startup time and when opening new files, and around 10-30x speedup for files that are already cached.

Further use of this method will be propogated to other methods in later diffs.

Reviewed By: aadisriram

Differential Revision: D5252102

fbshipit-source-id: 41e476a505a416854d9c76a7050883b39fda6b08