Solve the encoding issues again
Summary:
It seems to me that a previous issue was due to encoding. However, the encoding is fine and at some point someone passed an object to the log. I'm reverting to the original code and adding a check for non string objects.
It could be that a logged result from lldb is an object or something. So let's better catch the specific error and fix it properly.

Reviewed By: sas

Differential Revision: D5260990

fbshipit-source-id: b0fe704356d70db738ecebefbd99c8c2b9cedbf7