Fix working set changes modal not going away
Summary:
Previously not able to dismiss the working set selector modal by clicking out of it.  This was introduced by a change to make it so clicking 'cancel' during a delete confirmation wouldn't cause the modal to close.

The first time an atom.confirm dialog is created, an extra blur action is sent in which relatedTarget is not null (meaning clicking 'cancel' will still close the working set selector if it's the first confirm dialog that appeared since starting Nuclide).  All future atom.confirm dialogs will only generate blur actions with null relatedTarget.  I've demonstrated this in the test plan as well.

Reviewed By: shushz

Differential Revision: D5266914

fbshipit-source-id: e0541e86d54e6b976690be307bf5ac9a8bd485b3