Lock the height of Open Files when hovering File Tree
Summary:
When double clicking on a file in the File Tree, the newly opened file was being immediately added to Open Files section. This caused the Open Files section to become taller, thus shifting the File Tree down and shifting which file the user's mouse was hovering over. This can lead to great frustration when attempting to open several files in close succession.

The solution, as proposed by Christopher Chedeau, is to postpone updating the list of open files until the user's mouse has left the File Tree.

This implementation may be overly specific since theoretically we would like to lock the height of _all_ panels above the user's currently hovered panel. However, at the current time, File Tree is the only panel that has direct actions which can lead to an upper section being resized.

Feedback/criticism or ideas for alternative approaches are most welcome.

Reviewed By: matthewwithanm

Differential Revision:
D5176208

Tags: nuclide

fbshipit-source-id: b6d22ea8f0e07dd749a21cf28d827b42c5a8ef28