Move the buck hackery from the rpc to the client
Summary:
This was too gross.
I'm adding a new file that fetches in the client the path of the compilation database and passes it to the clang-rpc for it to do what is needed.

The main benefit of this change is that we will be able to get the buck flavors and modes when generating the database.

I'll remove the buck code in clang-rpc in a later diff. I made sure that it was not called when testing.

Something else I'll implement next is `reset` for these new providers. That will be trivial.

Reviewed By: hansonw

Differential Revision: D5254828

fbshipit-source-id: e014c75f6291be859107c7f9bb517ef50a162d3f