Improve the source inference
Summary:
Previously we were looking for source files related to a header file by strictly checking the basename and then getting the one closer to the header in the file system.
Now, besides doing that, we fallback to the closest source file in the file system, which works with the buck compilation databases.

Reviewed By: hansonw

Differential Revision: D5268771

fbshipit-source-id: 433c6045d1adf140bd17895eb5ea417491809003