Fix double echo in PHP debugger stdout
Summary: When launching a PHP script with the debugger attached, anything echo'ed to STDOUT is being printed to the console twice (once when we read it from stdout of the child process, and once when it's sent as a stdout message over the xdebug channel. Removing the callback from the child process create to eliminate the second output.

Reviewed By: mostafaeweda

Differential Revision: D5273376

fbshipit-source-id: 7f43792a589afce485e38fa12ad1f54d7fa02be8