Prevent tree-view's `finishLoading()` from being called twice
Summary:
This caused the resource path to be prepended to the package's path
multiple times, resulting in red boxes about stylesheet paths.
Unfortunately, we also need to clean up after tree-view's views which
may be left behind after its deactivation.

See also atom/atom#14837 and atom/tree-view#1136

This is nasty. We need to do something better, but this unblocks the release.

Reviewed By: jgebhardt

Differential Revision: D5281931

fbshipit-source-id: 2518a52648b00d5894124f3ac0f9d44ae04ad696