Have debugger share the workspace docks more effectively via nested panes
Summary:
This change addresses issues where the debugger did not share the dock properly with the Outline View or Commit form due to strange splitting of panes. In this change, I introduce a layout container that permits the debugger to nest panes inside panes.

With this, individual debugger panes can still be moved, split, resized and even torn out and moved to other docks, but they can also all live inside the same tab when sharing the dock with another full-height pane like Outline View or Commit form.

Reviewed By: matthewwithanm

Differential Revision: D5262415

fbshipit-source-id: 59f871d8d67f9be34821b6a8226d00203e419656