Introduce big-dig NPM module.
Summary:
`big-dig` extracts the guts of `nuclide-server` into its own reusable library
that is licensed with BSD+patents instead of the Nuclide license.
Unlike `nuclide-server`, there is no Python code: everything is in JavaScript.

Note that this revision does not update Nuclide itself to use `big-dig` instead
of `nuclide-server` yet, but that will be done in a future revision.

Reviewed By: hansonw

Differential Revision: D4512392

fbshipit-source-id: 364ab9755fe9c5b0c795c64e5a640b3a4ed25e66