*easy* autocomplete 'kind' and 'detail' fields
Summary:
The previous autocomplete implementation was a bit rough-and-ready. I'm starting to go through it carefully. This diff has two changes:

(1) The LSP 'kind' field indicates what icon should be used. This corresponds to the Atom 'type' field. This diff maps the two.

(2) The LSP 'detail' field contains a method-signature or similar, usually displayed at the top of the doc-comment-block for this suggestion. The equivalent place to display it in Atom is at the bottom of Autocomplete, in the 'description' field.

Reviewed By: arxanas

Differential Revision: D5302805

fbshipit-source-id: 7e6145a4fdc9f410639f01f703527de75260fd38