Support native platform group from native debugger package
Summary:
Currently we enable the debug button on the task runner toolbar for any target ending with "_binary" or "_test" - this results in it being enabled for targets that we actually don't know how to debug, and the user clicks to then see some cryptic error or embarassing failure.

The button should only be enabled opt-in for things we explicitly support.

This change has the Nuclide native debugger package offer a task runner platform group for Native that enables the supported actions.

Reviewed By: stepanhruda

Differential Revision: D5311552

fbshipit-source-id: b836781c041177fc9d14be77592acafa9b304917