Ensure that when build files are modified, clang and its providers' caches are cleared
Summary:
This makes the compilation db providers return a build file in case the compilation db is pending but a build file exists.

When a build file is provided to libclang.js, it will be watched and as soon as there are are changes in that build file, the caches will be cleared.

Reviewed By: hansonw

Differential Revision: D5314083

fbshipit-source-id: 093f7cb5ed881d72ff325417f52e755dc3af401d