Communicate code formatting failure details
Summary: I wanted to fix the gnarly formatting for Prettier errors and discovered that failure information is being communicated to the service via an Error (rejected Promise). I put together this quick diff but I think it actually shows that we should make the return type more robust (and capable of expressing failures). For now, though, this'll do.

Reviewed By: hansonw

Differential Revision: D5349249

fbshipit-source-id: 5511ce7f3a5060e050431f7a1824a069813f5459