Introduce DebuggerCapabilities and DebuggerProperties structure
Summary:
There are several features that are supported by some debuggers but not others (things like Threads, Run To Location, ...). Currently each subclass of DebugProcessInfo implements several supports*() rouintes that the debug infra has to invoke to determine if various things are supported or enabled.

Going forward, we plan to add several features that we want to be able to roll out to one debugger at a time, and with the current design this would end up requiring us to implement tons of supportsFoo() routines all over the place, which is messy and adds a bunch of extra cruft.

Instead, we will have each debugger return an object that describes everything it can (and can't do) in a single call.

This change introduces two methods on DebugProcessInfo:

getDebuggerCapabilities - reports features that a debugger supports / does not support.
getDebuggerProps - reports properties that a debugger can override (like the name or columns in the Threads window).

By contract, these items are all read-only to callers and do not change after the debug process info is created.

Reviewed By: mostafaeweda

Differential Revision: D5323363

fbshipit-source-id: 658a2751f32d574d090338f7f0b168b4281bdcc2