Fix react error when restarting PHP debugger
Summary:
When restarting the debugger using the button in the DebuggerControlsView, it ends up stopping and restarting the debugger rapidly in the same tick. React is throwing an error indicating that it is trying to process a deferred update batch for a DOM node that is now removed from the tree.

What I believe is happening is react is processing a deferred update for the DebuggerPaneContainer, but by the time it gets a chance to do that, we've already removed that container and added a new one in its place. The new nested pane container does more direct manipulation of the React DOM than the standard pane-based layout did, so we need to be more careful here.

This change fixes the invariant violation by stopping the debugger first, and then not starting it again until the next tick, which ensures that that happens after the deferred batch render. This appears to resolve the problem.

While I'm in here, also making a slight improvement to the "Restarting debugger..." notification - it was staying up too long, and it's positioning is especially painful because it's right over the default location of the debugger controls! Dismiss it automatically as soon as the debugger is started again.

Reviewed By: mostafaeweda

Differential Revision: D5339097

fbshipit-source-id: c6593d037f5807b9b28e5bdd58098dba522a5f80