Add a lint rule against shell-quote usages
Summary:
Until https://github.com/substack/node-shell-quote/pull/29, it's not a good idea to use `shell-quote` (This is exacerbated by the fact that many Facebook tools use paths with #).

This adds a lint warning against using it. I also realized that I need to update the path in `prefer-nuclide-uri` :D

Reviewed By: ebluestein

Differential Revision: D5354517

fbshipit-source-id: 157b9403d06f0804599e32af2705997db07b3165