Add jsx-a11y eslint plugin and disable failing rules
Summary:
This diff adds the jsx-a11y eslint plugin with its recommended settings and
disables all of the currently failing rules. Each is marked with an 'incremental' comment
so that if we deliberately decide to disable a particular rule these will be distinguishable.

We don't seem to accept recommended settings from other plugins, but in this case this plugin's
recommended settings are pretty specific with options passed to each rule and I think it's best
not to try to recreate that in our own config. Definitely open to suggestions though.

Reviewed By: jgebhardt

Differential Revision: D5270331

fbshipit-source-id: 579d78899da789ad60732b649026b8718fa952e5