Fixing active file monitoring in file-tree
Summary:
Since file-tree became a pane its active file tracking logic became flawed. It used to track pane changes and automatically scroll to an active file when such a change occured (if reveal on changes is enabled)

A single-click in the file tree would make it active pane and update the selected file. Unfortunately, however, this would also trigger the pane monitoring logic that would set the focus to an active file and effectively undo the file selection that has just been made.

This diff fixes the monitoring logic to not eagerly jump and select the selection to the current active file on each change, but rather only when the active file's path changes.

Reviewed By: matthewwithanm

Differential Revision: D5382079

fbshipit-source-id: 83ee1fe726605b9213c7f7b1471797fdedb7483b