Fix "Starting debugger opens file tree" bug
Summary:
The issue here is that during layout, we add a debugger pane container to both the left and right vertical docks, then add the debugger panes to their target docks, and then if one of the debugger pane containers is empty, we remove it.

In the case of the default layout, nothing gets added to the left pane's debugger dock, so it gets removed before layout completes (the user never sees it), but a side effect of the way it was being added is that the left dock pops open. So to the user, this looks like the file tree opens due to the debugger starting.

Ideally, we would just add the pane containers to the dock at the end, if and only if they aren't empty. Then we wouldn't have this problem. Unfortunately we have to hook them up at the beginning so that Atom attaches them to the DOM, even if briefly - this is necessary because if we don't, it's impossible for us to manipulate the flex sizes of the panes that get added, and we wouldn't be able to do things like collapse the amount of whitespace in the Debugger Controls pane (Atom would end up giving equal height to all the debugger panes, which is super lame).

So the fix here is to check up front if any pane will want to be spawned into the left or right dock pane container, and not add it if it's going to be empty.

Reviewed By: matthewwithanm

Differential Revision: D5358476

fbshipit-source-id: 0ebe73d4bd3eb0639512c1d2b8ec870308791a87