Support conditional breakpoints - Add configure condition command - 2/N
Summary:
This adds a "Configure breakpoint conditions..." command to the context menu for breakpoints in the gutter, and the context menu for breakpoints in the breakpoints debugger pane.

Clicking this option opens a modal dialog that prompts for a condition expression, and attaches that expression to the breakpoint.

Any breakpoint that has a condition now shows a question mark icon to the right of the bp in the breakpoints list, hovering over this icon shows a title tip with the condition, and clicking the icon opens the condition configuration dialog.

Reviewed By: mostafaeweda

Differential Revision: D5367945

fbshipit-source-id: aa73df79ad555f5f2060760c4509fe97bcac26b7