Add support for OCaml over the LSP
Summary:
This adds support for OCaml over the LSP, so that we can hopefully phase out our custom-built plugin and use an open-source one. That open-source one also has support for things like CodeLens to display the inferred type of a top-level function, which is something that I would really like to get in the future.

This is experimental, so I was thinking that ljw1004 and potentially others could try it out with me. It's not clear how well it performs under load; there might be upstream contributions needed to fix any performance issues.

Note that this currently requires that you install `ocaml-language-server` on your devserver manually. (It's a node package; I'm not sure about the process to get a node package installed on everyone's devserver.) You would also need to install `ocp-indent` on your devserver manually if you want to be able to format code (but this is the same for the existing OCaml plugin).

Reviewed By: ljw1004

Differential Revision: D5340647

fbshipit-source-id: d7d0d4f652a002bd632e591d4eb356170bd6afa3