Revert D5353592
Summary:
We are seeing a concerning number of crash logs where getVersion() is called for a buffer that has no BufferSubscription.  Because we do not yet understand the root cause, I am reverting this to avoid always returning 0 as the version of such buffers.

Note that other changes were made to this code after D5353592 (e.g. to use WeakMap) so this was not a fully automatic 'hg backout'.  Also, I had to re-introduce the Atom 1.18 changeCount property in atom.js.flow for this, which was not part of the original change.

Reviewed By: matthewwithanm

Differential Revision: D5412419

fbshipit-source-id: 3048dbca1680d2bc6d9967db0cb4e1dce064a015