Stop relying on changeCount, attempt 2
Summary:
The previous attempt had to be reverted because the fatal error was firing.

I think this is just happening when we're trying to get the version of the buffer after destruction - I was able to repro it by opening and closing a file very quickly, repeatedly, with the outline view open.

This brings back D5353592 and D5371606 with an extra `isDestroyed()` check. I want to get this in ASAP so we can let it soak for the entire week.

Reviewed By: matthewwithanm

Differential Revision: D5440680

fbshipit-source-id: 0ad163d0c9a86e47b2cf66cbc5ca76542f42e4d2