Add SocketService
Summary: Add an RPC socket service. The idea is to listen on a TCP socket on a remote host, and send down events when a client connects etc. In a further diff, I will react to the events, so when a client connects remotely, we will start a "mirror" client locally and connect to a specified local port + forward data back and forth, creating a tunnel.

Reviewed By: matthewwithanm

Differential Revision: D5382873

fbshipit-source-id: 4cdeb0988023278e0a2fd6e3609f5c94adbc3121