Fix warning messages poping up when enabling/disabling a breakpoint
Summary: The BreakpointManager for the new protocol channel is not handling disable correctly. It removes the breakpoint from the back-end but also from the internal list of breakpoints that it maintains inside the protocol channel. As a result, when you disable a bp and then re-enable it, a warning pops up that you have tried to enable a bp that doesn't exist. Then when you try to delete the bp, another warning pops up saying you're deleting a bp that doesn't exist. The state remains out of sync and broken.

Reviewed By: mostafaeweda

Differential Revision: D5421311

fbshipit-source-id: fcb7b96829ad0b4371d7d82b14262089a6a1e135