Support showing bp hit count in UX frontend
Summary:
This change adds a hit count field to breakpoints in the front end, which is undefined by default. If the hit count is defined and >= 0, a new green "Hit Count" label will appear beneath a breakpoint in the breakpoints pane. (See screenshot in test plan).  To be consistent, this also moves breakpoint Conditions beneath the BP using the same style.

NOTE: Nothing is attached to this yet, so it won't do anything as-is. This is just support in the front end, which will be used later when I add this to the debugger backends.

Reviewed By: mostafaeweda

Differential Revision: D5420565

fbshipit-source-id: 9a283dc8006aa6d722d03a0ec0321a49b1397598