Don't attempt to resolve bp in frontend if there is no location
Summary:
A user hit this issue when debugging an LLDB target with missing symbols. LLDB was unable to resolve the location of the breakpoint properly and unexpectedly sent the front end a NULL location, which causes this exception in Nuclide. If the backend sends us a bad location, we should be resiliant to that and not set the breakpoint in the UX.

The user's exception text:
Uncaught TypeError: Cannot match against 'undefined' or 'null'.

```
At /opt/facebook/nuclide/0.238.64761239/nuclide/node_modules/rxjs/bundles/Rx.min.js:43

TypeError: Cannot match against 'undefined' or 'null'.
  at BreakpointManager._createResolvedBreakpointFromLocation (/opt/facebook/nuclide/0.238.64761239/nuclide/pkg/nuclide-debugger/lib/Protocol/BreakpointManager.js:211:38)
  at BreakpointManager._sendBreakpointResolved (/opt/facebook/nuclide/0.238.64761239/nuclide/pkg/nuclide-debugger/lib/Protocol/BreakpointManager.js:202:51)
  at BreakpointManager.handleBreakpointResolved (/opt/facebook/nuclide/0.238.64761239/nuclide/pkg/nuclide-debugger/lib/Protocol/BreakpointManager.js:225:12)
  at BridgeAdapter._handleDebugEvent (/opt/facebook/nuclide/0.238.64761239/nuclide/pkg/nuclide-debugger/lib/Protocol/BridgeAdapter.js:212:35)

Reviewed By: mostafaeweda

Differential Revision: D5426945

fbshipit-source-id: 86310dd0d1762c39114f1dab943c7dac8561c217