Fix listing Java debuggable processes not working on Android O
Summary: In Android O they appear to have fixed the output of `ps`. On all prior versions, the data rows contained a column that contained an "S", but there was no corresponding header for that column. To work correctly, our parsing code had to skip that column when matching the data rows up with the column headers. As of Android O, there is a header "S" for this column.

Reviewed By: a20012251

Differential Revision: D5459821

fbshipit-source-id: 38e2cd6beea1676c7862a1dbd0d5bc35e1faa21e